refactor(identifiers): split personal.rs into per-region submodules
Convert `primitives/identifiers/common/patterns/personal.rs` (871 LOC)
into `personal/` with three per-region submodules:

- `us_identifiers.rs` (208 LOC): ssn, tax_id, driver_license, passport,
  employee_id, student_id
- `global_identifiers.rs` (250 LOC): national_id (UK NI / Canada SIN),
  korea_rrn, australia (tfn, abn), india (aadhaar, pan, gstin,
  vehicle_reg, voter_id, passport)
- `regional_identifiers.rs` (421 LOC): brazil (cpf, cnpj), mexico_curp,
  nigeria_nin, thailand_tnin, singapore_nric, finland_hetu, uk_ni,
  spain (nif, nie), italy_fiscal_code, poland_pesel, personal_name,
  birthdate

`personal/mod.rs` (39 LOC) re-exports all 28 sub-modules via
`pub(crate) use` so call sites continue to import e.g. `personal::ssn`
or `personal::india_pan` unchanged.

This is a move-only refactor — no logic, signature, visibility, or API
changes. Follows the precedent set by #317/#318/#331/#335.

Part of #338 (file-length umbrella). 1 of 7 PRs.
