--- crates/octarine/src/primitives/identifiers/common/patterns/personal.rs
@@ -1,871 +1,0 @@
-//! Shared regex patterns for identifier detection in text
-//!
-//! This module provides reusable regex patterns for scanning text documents
-//! to find various identifier types. These patterns are used by both detection
-//! and sanitization layers.
-//!
-#![allow(clippy::expect_used)]
-//! # Pattern Categories
-//!
-//! - **SSN**: Social Security Numbers (US)
-//! - **Tax IDs**: EIN, TIN, ITIN (US)
-//! - **Driver Licenses**: State-specific and generic patterns
-//! - **Passports**: International passport numbers
-//! - **Employee IDs**: Corporate identification
-//! - **Student IDs**: Educational identification
-//! - **National IDs**: International identification numbers
-//! - **Vehicle IDs**: VIN and license plates
-//!
-//! # Design Principles
-//!
-//! - **Conservative matching**: Prefer false negatives over false positives
-//! - **Context aware**: Use capture groups to preserve surrounding text
-//! - **Performance**: Use lazy_static for one-time compilation
-//! - **Extensibility**: Easy to add new patterns per identifier type
-
-// SAFETY: All regex patterns in this module are hardcoded and verified at compile time.
-// Regex::new() only fails on invalid syntax, which would be caught during development/testing.
-// Using expect() here is safe because these patterns are static and never change at runtime.
-
-use once_cell::sync::Lazy;
-use regex::Regex;
-use std::collections::HashMap;
-
-/// SSN (Social Security Number) patterns
-pub(crate) mod ssn {
-    use super::*;
-
-    /// SSN with explicit label (highest confidence)
-    /// Captures: prefix (label) + number
-    /// Example: "SSN: 123-45-6789" → groups: ("SSN: ", "123-45-6789")
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(\b(?i:SSN|Social[\s-]?Security)[\s:#-]*)(\d{3}-?\d{2}-?\d{4}\b)")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// SSN with dashes (high confidence)
-    /// Example: "123-45-6789"
-    pub static WITH_DASHES: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b\d{3}-\d{2}-\d{4}\b").expect("BUG: Invalid regex pattern"));
-
-    /// SSN with spaces (high confidence)
-    /// Example: "123 45 6789"
-    pub static WITH_SPACES: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b\d{3}\s\d{2}\s\d{4}\b").expect("BUG: Invalid regex pattern"));
-
-    /// SSN exact match pattern (for validation)
-    /// Example: "123-45-6789" or "123456789"
-    pub static EXACT: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"^\d{3}-?\d{2}-?\d{4}$").expect("BUG: Invalid regex pattern"));
-
-    /// Returns all SSN patterns in priority order
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*WITH_DASHES, &*WITH_SPACES]
-    }
-}
-
-/// Tax ID patterns (EIN, TIN, ITIN)
-pub(crate) mod tax_id {
-    use super::*;
-
-    /// Tax ID with explicit label (highest confidence)
-    /// Captures: prefix (label) + number
-    /// Example: "EIN: 12-3456789" → groups: ("EIN: ", "12-3456789")
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(\b(?:Company |)(?:TIN|EIN|ITIN)[\s#:-]+)(\d{2}-?\d{7}|\d{9})\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// EIN with dash format (high confidence)
-    /// Example: "12-3456789"
-    pub static EIN_FORMAT: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b\d{2}-\d{7}\b").expect("BUG: Invalid regex pattern"));
-
-    /// ITIN (starts with 9)
-    /// Example: "912-34-5678"
-    pub static ITIN_FORMAT: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b9\d{2}-?\d{2}-?\d{4}\b").expect("BUG: Invalid regex pattern"));
-
-    /// Generic TIN with prefix
-    pub static GENERIC_TIN: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(?:TIN|EIN|ITIN)[\s#:-]*\d{2}-?\d{7}\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// Federal EIN (FEIN)
-    pub static FEIN: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(?:FEIN|Federal[\s-]?EIN)[\s#:-]*\d{2}-?\d{7}\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![
-            &*LABELED,
-            &*EIN_FORMAT,
-            &*ITIN_FORMAT,
-            &*GENERIC_TIN,
-            &*FEIN,
-        ]
-    }
-}
-
-/// Driver's license patterns
-pub(crate) mod driver_license {
-    use super::*;
-
-    /// Generic driver's license pattern with context
-    /// Captures: prefix (label) + number
-    /// Example: "DL# A1234567" → groups: ("DL# ", "A1234567")
-    pub static GENERIC: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b((?i:DL|LICENSE|LIC)[\s#:-]*)([A-Z0-9]{6,15})\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// State-specific patterns (US)
-    pub fn state_patterns() -> HashMap<&'static str, Regex> {
-        let mut patterns = HashMap::new();
-        patterns.insert(
-            "CA",
-            Regex::new(r"\b[A-Z]\d{7}\b").expect("BUG: Invalid regex pattern"),
-        ); // California
-        patterns.insert(
-            "TX",
-            Regex::new(r"\b\d{8}\b").expect("BUG: Invalid regex pattern"),
-        ); // Texas
-        patterns.insert(
-            "NY",
-            Regex::new(r"\b[A-Z]\d{7}|[A-Z]\d{18}\b").expect("BUG: Invalid regex pattern"),
-        ); // New York
-        patterns.insert(
-            "FL",
-            Regex::new(r"\b[A-Z]\d{12}\b").expect("BUG: Invalid regex pattern"),
-        ); // Florida
-        patterns
-    }
-}
-
-/// Passport patterns
-pub(crate) mod passport {
-    use super::*;
-
-    /// Explicit passport mention (highest confidence)
-    /// Examples: "Passport: 123456789", "Passport number: 123456789", "Passport No: 123456789"
-    pub static EXPLICIT: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(?i:passport)(?:\s+(?:number|no|num|#))?[\s#:-]+\d{8,9}\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Passport with PP prefix
-    /// Captures: prefix + number
-    /// Example: "PP# 987654321" → groups: ("PP# ", "987654321")
-    pub static WITH_PREFIX: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b((?i:pp)[\s#:-]*)(\d{9})\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// Generic passport format (letter + digits)
-    /// Example: "C12345678"
-    /// Note: Lower confidence, needs context checking
-    pub static GENERIC: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b([A-Z]\d{6,9})\b").expect("BUG: Invalid regex pattern"));
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*EXPLICIT, &*WITH_PREFIX, &*GENERIC]
-    }
-}
-
-/// Employee ID patterns
-pub(crate) mod employee_id {
-    use super::*;
-
-    /// Employee ID with explicit prefix
-    /// Example: "EMP00123", "Employee: 12345"
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(?:EMP|EMPLOYEE|STAFF)[\s#:-]*[A-Z0-9]{4,12}\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Badge number
-    /// Example: "BADGE# 98765"
-    pub static BADGE: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(?:BADGE|ID)[\s#:-]*\d{4,10}\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// E-number format (common in corporations)
-    /// Example: "E123456"
-    pub static E_NUMBER: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b[Ee]\d{5,8}\b").expect("BUG: Invalid regex pattern"));
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*BADGE, &*E_NUMBER]
-    }
-}
-
-/// Student ID patterns
-pub(crate) mod student_id {
-    use super::*;
-
-    /// Student ID with explicit prefix
-    /// Example: "STUDENT# 1234567"
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(?:STUDENT|STU|STUD)[\s#:-]*[A-Z0-9]{6,12}\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// S-number format (common in universities)
-    /// Example: "S12345678"
-    pub static S_NUMBER: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b[Ss]\d{7,9}\b").expect("BUG: Invalid regex pattern"));
-
-    /// Formatted with dashes
-    /// Example: "123-45-6789"
-    /// Note: Overlaps with SSN format, needs context
-    pub static WITH_DASHES: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b\d{3}-\d{2}-\d{4}\b").expect("BUG: Invalid regex pattern"));
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*S_NUMBER, &*WITH_DASHES]
-    }
-}
-
-/// National ID patterns (international)
-pub(crate) mod national_id {
-    use super::*;
-
-    /// UK National Insurance Number
-    /// Example: "AB123456C"
-    pub static UK_NI: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b[A-Z]{2}\d{6}[A-Z]\b").expect("BUG: Invalid regex pattern"));
-
-    /// Canadian Social Insurance Number (SIN)
-    /// Example: "123-456-789" or "123 456 789"
-    pub static CANADA_SIN: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b\d{3}[-\s]?\d{3}[-\s]?\d{3}\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// Generic national ID with label
-    /// Example: "NATIONAL ID: ABC123456789"
-    pub static GENERIC: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(?:NATIONAL[\s-]?ID|NID)[\s#:-]*[A-Z0-9]{8,15}\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*UK_NI, &*CANADA_SIN, &*GENERIC]
-    }
-}
-/// South Korea Resident Registration Number patterns
-pub(crate) mod korea_rrn {
-    use super::*;
-
-    /// Korea RRN with dash: YYMMDD-GNNNNNN (13 digits total)
-    /// Gender digit 1-8 follows the dash
-    pub static WITH_DASH: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b\d{6}-[1-8]\d{6}\b").expect("BUG: Invalid regex pattern"));
-
-    /// Korea RRN with explicit label (with or without dash)
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:RRN|주민등록번호|resident[\s-]?registration)[\s:#-]*(\d{6}-?[1-8]\d{6})\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*WITH_DASH]
-    }
-}
-
-/// Australian Tax File Number patterns
-pub(crate) mod australia_tfn {
-    use super::*;
-
-    /// TFN with spaces: NNN NNN NNN (9 digits)
-    pub static WITH_SPACES: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b\d{3}\s\d{3}\s\d{3}\b").expect("BUG: Invalid regex pattern"));
-
-    /// TFN with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(?i)\b(?:TFN|tax[\s-]?file[\s-]?number)[\s:#-]*(\d{3}\s?\d{3}\s?\d{3})\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*WITH_SPACES]
-    }
-}
-
-/// Australian Business Number patterns
-pub(crate) mod australia_abn {
-    use super::*;
-
-    /// ABN with spaces: NN NNN NNN NNN (11 digits)
-    pub static WITH_SPACES: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b\d{2}\s\d{3}\s\d{3}\s\d{3}\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// ABN with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:ABN|australian[\s-]?business[\s-]?number)[\s:#-]*(\d{2}\s?\d{3}\s?\d{3}\s?\d{3})\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*WITH_SPACES]
-    }
-}
-
-/// India Aadhaar number patterns (12 digits, starts with 2-9)
-pub(crate) mod india_aadhaar {
-    use super::*;
-
-    /// Aadhaar with spaces: NNNN NNNN NNNN (starts with 2-9)
-    pub static WITH_SPACES: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b[2-9]\d{3}\s\d{4}\s\d{4}\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// Aadhaar with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(?i)\b(?:aadhaar|aadhar|UIDAI)[\s:#-]*([2-9]\d{3}\s?\d{4}\s?\d{4})\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*WITH_SPACES]
-    }
-}
-
-/// India PAN (Permanent Account Number) patterns
-pub(crate) mod india_pan {
-    use super::*;
-
-    /// PAN format: AAAAA9999A (5 letters + 4 digits + 1 letter)
-    pub static STANDARD: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b[A-Z]{5}\d{4}[A-Z]\b").expect("BUG: Invalid regex pattern"));
-
-    /// PAN with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:PAN|permanent[\s-]?account[\s-]?number)[\s:#-]*([A-Z]{5}\d{4}[A-Z])\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-/// India GSTIN (Goods and Services Tax Identification Number) patterns
-///
-/// Format: 15 chars = 2-digit state code + 10-char PAN + entity number (1-9 or A-Z)
-/// + literal 'Z' + check character (alphanumeric)
-pub(crate) mod india_gstin {
-    use super::*;
-
-    /// GSTIN standard format: NN AAAAA NNNN A [1-9A-Z] Z [0-9A-Z]
-    pub static STANDARD: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b\d{2}[A-Z]{5}\d{4}[A-Z][1-9A-Z]Z[0-9A-Z]\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// GSTIN with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:GSTIN|GST[\s-]?(?:identification)?[\s-]?(?:number|no\.?))[\s:#-]*(\d{2}[A-Z]{5}\d{4}[A-Z][1-9A-Z]Z[0-9A-Z])\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-/// India Vehicle Registration (license plate) patterns
-///
-/// Format: state code (2 letters) + district code (1-2 digits) + optional series
-/// (0-3 letters) + number (1-4 digits). Examples: MH02AB1234, DL1C1234, KA01MA1234.
-pub(crate) mod india_vehicle_reg {
-    use super::*;
-
-    /// Vehicle registration standard format (no spaces)
-    pub static STANDARD: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b[A-Z]{2}\d{1,2}[A-Z]{1,3}\d{1,4}\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// Vehicle registration with spaces or hyphens between segments
-    pub static WITH_SEPARATORS: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b[A-Z]{2}[\s-]\d{1,2}[\s-][A-Z]{1,3}[\s-]\d{1,4}\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Vehicle registration with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:vehicle[\s-]?(?:registration|number)|license[\s-]?plate|reg[\s-]?no\.?)[\s:#-]*([A-Z]{2}[\s-]?\d{1,2}[\s-]?[A-Z]{1,3}[\s-]?\d{1,4})\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*WITH_SEPARATORS, &*STANDARD]
-    }
-}
-
-/// India Voter ID (EPIC - Electors Photo Identity Card) patterns
-///
-/// Format: 3 letters (state/constituency code) + 7 digits
-pub(crate) mod india_voter_id {
-    use super::*;
-
-    /// Voter ID standard format
-    pub static STANDARD: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b[A-Z]{3}\d{7}\b").expect("BUG: Invalid regex pattern"));
-
-    /// Voter ID with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:voter[\s-]?(?:ID|card|EPIC)?|EPIC|elector(?:'?s)?[\s-]?photo[\s-]?identity[\s-]?card)[\s:#-]*([A-Z]{3}\d{7})\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-/// Indian Passport patterns
-///
-/// Format: 1 letter (type indicator: P=personal, S=service, D=diplomatic) + 7 digits.
-/// STANDARD pattern is short and prone to false positives — finders use LABELED
-/// only, while direct `is_india_passport()` checks accept STANDARD.
-pub(crate) mod india_passport {
-    use super::*;
-
-    /// Indian Passport standard format
-    pub static STANDARD: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b[A-Z]\d{7}\b").expect("BUG: Invalid regex pattern"));
-
-    /// Indian Passport with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:indian[\s-]?passport|passport[\s-]?(?:no\.?|number)?|IN[\s-]?passport)[\s:#-]*([A-Z]\d{7})\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Returns patterns used for text scanning (LABELED only — STANDARD is
-    /// `[A-Z]\d{7}` which would match many non-passport strings).
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED]
-    }
-}
-
-/// Brazil CPF (Cadastro de Pessoas Físicas) patterns
-///
-/// Format: `NNN.NNN.NNN-NN` or 11 plain digits.
-pub(crate) mod brazil_cpf {
-    use super::*;
-
-    /// Formatted CPF: NNN.NNN.NNN-NN
-    pub static FORMATTED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b\d{3}\.\d{3}\.\d{3}-\d{2}\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// CPF with explicit label (accepts formatted or unformatted)
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:CPF|cadastro[\s-]?de[\s-]?pessoas[\s-]?f[ií]sicas)[\s:#-]*(\d{3}\.?\d{3}\.?\d{3}-?\d{2})\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Scanning patterns. Plain 11-digit form is too ambiguous (matches phone
-    /// numbers, IDs, etc.) so we only scan formatted or labeled occurrences.
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*FORMATTED]
-    }
-}
-
-/// Brazil CNPJ (Cadastro Nacional da Pessoa Jurídica) patterns
-///
-/// Format: `NN.NNN.NNN/NNNN-NN` or 14 plain digits.
-pub(crate) mod brazil_cnpj {
-    use super::*;
-
-    /// Formatted CNPJ: NN.NNN.NNN/NNNN-NN
-    pub static FORMATTED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b\d{2}\.\d{3}\.\d{3}/\d{4}-\d{2}\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// CNPJ with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:CNPJ|cadastro[\s-]?nacional)[\s:#-]*(\d{2}\.?\d{3}\.?\d{3}/?\d{4}-?\d{2})\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*FORMATTED]
-    }
-}
-
-/// Mexico CURP (Clave Única de Registro de Población) patterns
-///
-/// Format: 4 letters + 6 digits (YYMMDD) + gender (H/M) + 2 letters (state) +
-/// 3 letters + alphanumeric + digit.
-pub(crate) mod mexico_curp {
-    use super::*;
-
-    /// CURP standard format (18 characters)
-    pub static STANDARD: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b[A-Z]{4}\d{6}[HM][A-Z]{5}[0-9A-Z]\d\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// CURP with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:CURP|clave[\s-]?[uú]nica[\s-]?de[\s-]?registro)[\s:#-]*([A-Z]{4}\d{6}[HM][A-Z]{5}[0-9A-Z]\d)\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-/// Nigeria NIN (National Identification Number) patterns
-///
-/// Format: 11 plain digits. The unlabeled form is identical to phone numbers
-/// and many other identifiers, so scanning uses LABELED only — analogous to
-/// `india_passport`. Direct `is_nigeria_nin()` checks still accept the bare
-/// 11-digit form.
-pub(crate) mod nigeria_nin {
-    use super::*;
-
-    /// NIN standard format (11 digits, no separators)
-    pub static STANDARD: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b\d{11}\b").expect("BUG: Invalid regex pattern"));
-
-    /// NIN with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(?i)\b(?:NIN|national[\s-]?identification[\s-]?number|NIMC)[\s:#-]*(\d{11})\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Returns patterns used for text scanning (LABELED only — STANDARD is
-    /// `\d{11}` which matches phone numbers and many other 11-digit strings).
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED]
-    }
-}
-
-/// Thailand TNIN (National Identification Number) patterns
-///
-/// Format: 13 digits. Display form `N-NNNN-NNNNN-NN-N`.
-pub(crate) mod thailand_tnin {
-    use super::*;
-
-    /// Formatted TNIN: N-NNNN-NNNNN-NN-N
-    pub static FORMATTED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b\d-\d{4}-\d{5}-\d{2}-\d\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// TNIN with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:TNIN|Thai[\s-]?national[\s-]?ID|TID)[\s:#-]*(\d-?\d{4}-?\d{5}-?\d{2}-?\d)\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Scanning patterns. Plain 13-digit form is too ambiguous, so only
-    /// formatted or labeled occurrences are scanned.
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*FORMATTED]
-    }
-}
-
-/// Singapore NRIC/FIN patterns
-pub(crate) mod singapore_nric {
-    use super::*;
-
-    /// NRIC/FIN format: [STFGM] + 7 digits + check letter
-    pub static STANDARD: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b[STFGM]\d{7}[A-Z]\b").expect("BUG: Invalid regex pattern"));
-
-    /// NRIC/FIN with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(?i)\b(?:NRIC|FIN|identity[\s-]?card)[\s:#-]*([STFGM]\d{7}[A-Z])\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-/// Finland HETU (personal identity code) patterns
-pub(crate) mod finland_hetu {
-    use super::*;
-
-    /// HETU format: DDMMYY[+-A]NNNC (6 digits + century + 3 digits + check)
-    pub static STANDARD: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b\d{6}[-+A]\d{3}[0-9A-Y]\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// HETU with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:HETU|henkilotunnus|personal[\s-]?identity[\s-]?code)[\s:#-]*(\d{6}[-+A]\d{3}[0-9A-Y])\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-/// UK National Insurance Number (NINO) patterns
-///
-/// Format: 2 letters + 6 digits + 1 suffix letter, e.g. `AB123456C`.
-/// The `is_uk_ni` / `find_uk_nis_in_text` detection functions apply HMRC
-/// prefix/suffix validation (BG, GB, NK, KN, TN, NT, ZZ excluded; suffix
-/// must be A-D) on top of these shape-only patterns.
-pub(crate) mod uk_ni {
-    use super::*;
-
-    /// Bare UK NINO shape: 2 letters + 6 digits + 1 letter
-    /// Example: `AB123456C`
-    pub static STANDARD: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(?i)\b[A-Z]{2}\d{6}[A-Z]\b").expect("BUG: Invalid regex pattern")
-    });
-
-    /// UK NINO with explicit label (two capture groups: label prefix, NINO value)
-    /// Example: `NI: AB123456C`, `NINO AB123456C`, `National Insurance: AB123456C`
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(\b(?i:NINO|NI(?:[\s-]?Number)?|National[\s-]?Insurance(?:[\s-]?Number)?)[\s:#-]*)((?i:[A-Z]{2}\d{6}[A-Z]))\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-/// Spain NIF (Numero de Identificacion Fiscal) patterns
-pub(crate) mod spain_nif {
-    use super::*;
-
-    /// NIF format: 8 digits + 1 check letter
-    pub static STANDARD: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"(?i)\b\d{8}[A-Z]\b").expect("BUG: Invalid regex pattern"));
-
-    /// NIF with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(?i)\b(?:NIF|DNI|documento[\s-]?nacional)[\s:#-]*(\d{8}[A-Z])\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-/// Spain NIE (Numero de Identidad de Extranjero) patterns
-pub(crate) mod spain_nie {
-    use super::*;
-
-    /// NIE format: X/Y/Z + 7 digits + 1 check letter
-    pub static STANDARD: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"(?i)\b[XYZ]\d{7}[A-Z]\b").expect("BUG: Invalid regex pattern"));
-
-    /// NIE with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(?i)\b(?:NIE|numero[\s-]?de[\s-]?identidad[\s-]?de[\s-]?extranjero)[\s:#-]*([XYZ]\d{7}[A-Z])\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-/// Italy Codice Fiscale (fiscal code) patterns
-pub(crate) mod italy_fiscal_code {
-    use super::*;
-
-    /// Codice Fiscale format: 6 letters + 2 digits + 1 letter + 2 digits + 1 letter + 3 digits + 1 letter
-    pub static STANDARD: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(?i)\b[A-Z]{6}\d{2}[A-Z]\d{2}[A-Z]\d{3}[A-Z]\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Codice Fiscale with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:codice[\s-]?fiscale|C\.?F\.?|fiscal[\s-]?code)[\s:#-]*([A-Z]{6}\d{2}[A-Z]\d{2}[A-Z]\d{3}[A-Z])\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-/// Poland PESEL (personal identity number) patterns
-pub(crate) mod poland_pesel {
-    use super::*;
-
-    /// PESEL format: 11 consecutive digits (YYMMDDNNNCC)
-    pub static STANDARD: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"\b\d{11}\b").expect("BUG: Invalid regex pattern"));
-
-    /// PESEL with explicit label
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i)\b(?:PESEL|numer[\s-]?PESEL|personal[\s-]?id(?:entity)?[\s-]?(?:number|code))[\s:#-]*(\d{11})\b",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*STANDARD]
-    }
-}
-
-pub(crate) mod personal_name {
-    use super::*;
-
-    /// First Last (with optional middle initial/name)
-    /// Example: "John Smith", "Mary Jane Doe", "Bob A. Wilson"
-    /// **Warning**: High false positive rate - matches any capitalized words
-    pub static FIRST_LAST: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b[A-Z][a-z]+(?:\s+[A-Z]\.?)?\s+[A-Z][a-z]+\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Last, First (with optional middle)
-    /// Example: "Smith, John", "Doe, Mary Jane"
-    pub static LAST_FIRST: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b[A-Z][a-z]+,\s*[A-Z][a-z]+(?:\s+[A-Z]\.?)?\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Labeled name
-    /// Example: "Name: John Smith", "Patient: Mary Doe"
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"(?i:name|patient|client|customer)[\s:]+([A-Z][a-z]+\s+[A-Z][a-z]+)")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![&*LABELED, &*LAST_FIRST, &*FIRST_LAST]
-    }
-}
-
-/// Birthdate and date patterns
-///
-/// Supports multiple date formats commonly used for birthdates.
-pub(crate) mod birthdate {
-    use super::*;
-
-    /// ISO format: YYYY-MM-DD
-    /// Example: "1990-05-15", "2000-12-31"
-    pub static ISO_FORMAT: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(19|20)\d{2}[-.](0[1-9]|1[0-2])[-.](0[1-9]|[12]\d|3[01])\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// US format: MM/DD/YYYY or MM-DD-YYYY
-    /// Example: "05/15/1990", "12-31-2000"
-    pub static US_FORMAT: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(0[1-9]|1[0-2])[/\-](0[1-9]|[12]\d|3[01])[/\-](19|20)\d{2}\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// European format: DD/MM/YYYY or DD-MM-YYYY
-    /// Example: "15/05/1990", "31-12-2000"
-    pub static EU_FORMAT: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(0[1-9]|[12]\d|3[01])[/\-](0[1-9]|1[0-2])[/\-](19|20)\d{2}\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Month name format: "Month DD, YYYY"
-    /// Example: "January 15, 1990", "Dec 31, 2000"
-    pub static MONTH_NAME: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?x)
-            \b(January|February|March|April|May|June|July|August|September|October|November|December|
-               Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)
-            \s+(0?[1-9]|[12]\d|3[01]),?\s+(19|20)\d{2}\b
-            ",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Day-Month-Year with abbreviated month name
-    /// Example: "15-Jan-1990", "15 Jan 1990", "1 Feb 2000"
-    pub static DAY_MONTH_ABBREV: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?x)
-            \b(0?[1-9]|[12]\d|3[01])[-.\s]
-            (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)[-.\s]?
-            (19|20)\d{2}\b
-            ",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Year-first with slashes: YYYY/MM/DD
-    /// Example: "1990/01/15", "2000/12/31"
-    pub static YEAR_FIRST_SLASH: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(19|20)\d{2}/(0[1-9]|1[0-2])/(0[1-9]|[12]\d|3[01])\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Two-digit year (US-style): MM/DD/YY or MM-DD-YY
-    /// Example: "01/15/90", "12-31-00"
-    pub static TWO_DIGIT_YEAR: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(0[1-9]|1[0-2])[/\-](0[1-9]|[12]\d|3[01])[/\-]\d{2}\b")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// ISO format with time portion (date part extracted)
-    /// Example: "1990-01-15T10:30:00", "2023-06-15T08:00:00Z"
-    pub static ISO_WITH_TIME: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(r"\b(19|20)\d{2}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])T\d{2}:\d{2}")
-            .expect("BUG: Invalid regex pattern")
-    });
-
-    /// Labeled birthdate
-    /// Example: "DOB: 1990-05-15", "Born: May 15, 1990"
-    pub static LABELED: Lazy<Regex> = Lazy::new(|| {
-        Regex::new(
-            r"(?i:dob|birthdate|born|birth)[\s:]+([0-9/\-]{8,10}|[A-Za-z]+\s+\d{1,2},?\s+\d{4})",
-        )
-        .expect("BUG: Invalid regex pattern")
-    });
-
-    pub fn all() -> Vec<&'static Regex> {
-        vec![
-            &*LABELED,
-            &*MONTH_NAME,
-            &*DAY_MONTH_ABBREV,
-            &*ISO_WITH_TIME,
-            &*ISO_FORMAT,
-            &*YEAR_FIRST_SLASH,
-            &*US_FORMAT,
-            &*EU_FORMAT,
-            &*TWO_DIGIT_YEAR,
-        ]
-    }
-}
